% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_xsyg2xlum.R
\name{convert_xsyg2xlum}
\alias{convert_xsyg2xlum}
\title{Convert Freiberg Instruments XSYG files to XLUM}
\usage{
convert_xsyg2xlum(file, out_file = NULL)
}
\arguments{
\item{file}{\link{character} (\strong{required}): file}

\item{out_file}{\link{character} (\emph{optional}): output file, if set a file
output is created}
}
\value{
Depending on the setting. If \code{out_file == NULL} an \link[xml2:xml_document-class]{xml2::xml_document} object is returned,
if file is set the function attempts to write an \verb{*.xlum} file`.
}
\description{
Wrapped function to quickly convert Freiberg Instruments XSYG-files
to \code{xlum} (files)
}
\section{Function version}{
 0.1.0
}

\examples{
## load package example data
## for loading own data use
## file <- file.choose()
file <- system.file("extdata/XSYG_file.xsyg", package = "Luminescence")
convert_xsyg2xlum(file)

}
\seealso{
\link{convert_rlum2xlum}, \link[Luminescence:read_XSYG2R]{Luminescence::read_XSYG2R}
}
\author{
Sebastian Kreutzer, Institute of Geography, Heidelberg University, Heidelberg (Germany)
}
