% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_xlum.R
\name{validate_xlum}
\alias{validate_xlum}
\title{Validate xlum format against an XSD reference schema}
\usage{
validate_xlum(file, xsd_version = NULL)
}
\arguments{
\item{file}{\link[xml2:xml_document-class]{xml2::xml_document} (\strong{required}): object to test}

\item{xsd_version}{\link{numeric} (\emph{optional}): specify format version for the
validation, e.g., \code{1}, \code{1.1}. The default is \code{NULL}, which will use the most recent version}
}
\value{
Results of the validation \link{logical} with attributes
}
\description{
A a convenience wrapper around \link[xml2:xml_validate]{xml2::xml_validate} against
the reference format description shipped with \link{xlum-package}
}
\examples{
file <- system.file("extdata/xlum_prototype.xlum", package = "xlum")
validate_xlum(file)

}
\author{
Sebastian Kreutzer, Geography & Earth Sciences, Aberystwyth University
}
