% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_binx2xlum.R
\name{convert_binx2xlum}
\alias{convert_binx2xlum}
\alias{convert_bin2xlum}
\title{Convert Risø BIN/BINX to XLUM}
\usage{
convert_binx2xlum(file, out_file = NULL)

convert_bin2xlum(file, out_file = NULL)
}
\arguments{
\item{file}{\link{character} (\strong{required}): file}

\item{out_file}{\link{character} (\emph{optional}): output file, if set a file
output is created}
}
\value{
Depending on the setting. If \code{out_file == NULL} an \link[xml2:oldclass]{xml2::xml_document} object is returned,
if file is set the function attempts to write an \verb{*.xlum} file`.
}
\description{
Convert Risø BIN/BINX to XLUM
}
\section{Function version}{
 0.1.0
}

\examples{

\dontrun{
file <- system.file("extdata/BINfile_V8.binx", package = "Luminescence")
convert_binx2xlum(file)
}

}
\seealso{
\link{convert_rlum2xlum}, \link[Luminescence:read_BIN2R]{Luminescence::read_BIN2R}
}
\author{
Sebastian Kreutzer, Geography & Earth Sciences, Aberystwyth University (United Kingdom)
} 

\section{How to cite}{
Kreutzer, S., 2022. convert_binx2xlum(): Convert Risø BIN/BINX to XLUM. Function version 0.1.0. In: Kreutzer, S., 2022. xlum: Read, Write, and Convert XLUM Data. R package version 0.1.0.9000-76. 
}

