% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_psl2xlum.R
\name{convert_psl2xlum}
\alias{convert_psl2xlum}
\title{Convert SUERC PSL-files to XLUM}
\usage{
convert_psl2xlum(file, out_file = NULL)
}
\arguments{
\item{file}{\link{character} (\strong{required}): file}

\item{out_file}{\link{character} (\emph{optional}): output file, if set a file
output is created}
}
\value{
Depending on the setting. If \code{out_file == NULL} an \link[xml2:oldclass]{xml2::xml_document} object is returned,
if file is set the function attempts to write an \verb{*.xlum} file`.
}
\description{
Wrapped function to quickly convert PSL files
to \code{xlum} (files) from the portable luminescence system produced by the SUERC
}
\section{Function version}{
 0.1.0
}

\examples{
## load package example data
## for loading own data use
## file <- file.choose()
file <- system.file("extdata", "DorNie_0016.psl", package = "Luminescence")
convert_psl2xlum(file)

}
\seealso{
\link{convert_rlum2xlum}, \link[Luminescence:read_PSL2R]{Luminescence::read_PSL2R}
}
\author{
Sebastian Kreutzer, Institute of Geography, Heidelberg University, Heidelberg (Germany)
} 

\section{How to cite}{
Kreutzer, S., 2022. convert_psl2xlum(): Convert SUERC PSL-files to XLUM. Function version 0.1.0. In: Kreutzer, S., 2022. xlum: Read, Write, and Convert XLUM Data. R package version 0.1.0.9000-91. 
}

