% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_rlum2xlum.R
\name{convert_rlum2xlum}
\alias{convert_rlum2xlum}
\title{Convert RLum-class objects to the XLUM format}
\usage{
convert_rlum2xlum(rlum, file = NULL)
}
\arguments{
\item{rlum}{\link[Luminescence:RLum.Data-class]{Luminescence::RLum.Data} and \link[Luminescence:RLum.Analysis-class]{Luminescence::RLum.Analysis} or a \link{list} of it (\strong{required}): input
for the conversion}

\item{file}{\link{character} (\emph{optional}): file name for the export, if \code{NULL} (the default),
an \link[xml2:oldclass]{xml2::xml_document} object is returned}
}
\value{
Depending on the setting. If \code{file == NULL} an \link[xml2:oldclass]{xml2::xml_document} object is returned,
if file is set the function attempts to write an \verb{*.xlum} file`.
}
\description{
Converts \link[Luminescence:RLum-class]{Luminescence::RLum} objects to xlum format
}
\details{
The function tries to make a best possible conversion of the \link[Luminescence:RLum-class]{Luminescence::RLum}
to the \code{xlum} format. \link[Luminescence:RLum.Results-class]{Luminescence::RLum.Results} objects are not supported.

Because of the nature the \link[Luminescence:RLum-class]{Luminescence::RLum} structure, the nodes \verb{<xlum/>}, \verb{<sample/>}
are set automatically because the \link[Luminescence:RLum-class]{Luminescence::RLum} does not distinguish between
different samples.
}
\section{Function version}{
 0.1.0
}

\examples{
data(ExampleData.RLum.Analysis,
envir = environment(),
package = "Luminescence")
convert_rlum2xlum(IRSAR.RF.Data)

}
\seealso{
\link[Luminescence:RLum.Data-class]{Luminescence::RLum.Data}, \link{write_xlum}
}
\author{
Sebastian Kreutzer, Geography & Earth Sciences, Aberystwyth University (United Kingdom)
} 

\section{How to cite}{
Kreutzer, S., 2022. convert_rlum2xlum(): Convert RLum-class objects to the XLUM format. Function version 0.1.0. In: Kreutzer, S., 2022. xlum: Read, Write, and Convert XLUM Data. R package version 0.1.0.9000-75. 
}

\keyword{datagen}
