% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_xlum.R
\name{read_xlum}
\alias{read_xlum}
\title{Import xlum files}
\usage{
read_xlum(file, verify = TRUE, output = "xlum_list")
}
\arguments{
\item{file}{\link{character} (\strong{required}): path to file}

\item{verify}{\link{logical} (\emph{with default}): enable/disable format validation}

\item{output}{\link{character} (\emph{with default}): output object of the import,
supported are \code{"xml_document"}, \code{"list"}, \code{"xlum_list"} (the default)}
}
\value{
##TODO
}
\description{
#TODO
}
\details{
\strong{Supported output options}\cr

\code{"xlum_list"} (default): imports XML-data and coerces them to list. Numeric values
for level \verb{<curve>} are transposed into \link{numeric} vectors for better handling

\code{"list"}: imports XML-data and coerces it to a \link{list}, no further treatment of data

\code{"xml_document"}: unprocessed output from \link[xml2:read_xml]{xml2::read_xml}
}
\author{
Sebastian Kreutzer, Geography & Earth Sciences, Aberystwyth University
} 

\section{How to cite}{
Kreutzer, S., 2022. read_xlum(): Import xlum files. In: Kreutzer, S., 2022. xlum: Read, Write, and Convert xlum Data. R package version 0.1.0.9000-8. 
}

