% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_xlum.R
\name{validate_xlum}
\alias{validate_xlum}
\title{Validate xlum format against an XSD reference schema}
\usage{
validate_xlum(file)
}
\arguments{
\item{file}{\link[xml2:oldclass]{xml2::xml_document} (\strong{required}): object to test}
}
\value{
Results of the validation \link{logical} with attributes
}
\description{
A a convenience wrapper around \link[xml2:xml_validate]{xml2::xml_validate} against
the reference format description shipped with \link{xlum-package}
}
\examples{
file <- system.file("extdata/xlum_prototype.xlum", package = "xlum")
validate_xlum(file)

}
\author{
Sebastian Kreutzer, Geography & Earth Sciences, Aberystwyth University
} 

\section{How to cite}{
Kreutzer, S., 2022. validate_xlum(): Validate xlum format against an XSD reference schema. In: Kreutzer, S., 2022. xlum: Read, Write, and Convert xlum Data. R package version 0.1.0.9000-13. 
}

