% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_xlum.R
\name{write_xlum}
\alias{write_xlum}
\title{Write XLUM-files}
\usage{
write_xlum(x, file, ...)
}
\arguments{
\item{x}{\code{xlum_list}, \link{list}, \link[xml2:oldclass]{xml2::xml_document} (\strong{required}) : input object}

\item{file}{\link{character} (\strong{required}): valid file path and file name with ending
\code{.xlum} (will be add automatically if needed)}

\item{...}{further arguments to be passed to \link[xml2:write_xml]{xml2::write_xml}}
}
\value{
Creates an XML-file with ending \verb{*.xlum}
}
\description{
Create XLUM-files using \link[xml2:read_xml]{xml2::read_xml}
}
\section{Function version}{
 0.1.0
}

\examples{
file <- system.file("extdata/xlum_prototype.xlum", package="xlum")
out_file <- tempfile()

xlum_data <- read_xlum(file)
write_xlum(xlum_data, out_file)

}
\seealso{
\link[xml2:write_xml]{xml2::write_xml}, \link{read_xlum}
}
\author{
Sebastian Kreutzer, Geography & Earth Sciences, Aberystwyth University (united Kingdom)
} 

\section{How to cite}{
Kreutzer, S., 2022. write_xlum(): Write XLUM-files. Function version 0.1.0. In: Kreutzer, S., 2022. xlum: Read, Write, and Convert XLUM Data. R package version 0.1.0.9000-64. 
}

\keyword{IO}
