% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlum-package.R
\docType{package}
\name{xlum-package}
\alias{xlum-package}
\alias{xlum}
\title{Read, Write, and Convert xlum Data\cr

\if{html}{
\figure{xlum_logo.png}{options: width="75"}
}}
\description{
Support for the XLUM-file format for exchange and long-term
preservation of luminescence data. Facilitated are the import and export
of XLUM-files and the conversion from other ASCII and binary formats commonly
used to store luminescence (dating) data.
}
\details{
\strong{Support contact}
\itemize{
\item \url{https://github.com/R-Lum/xlum/discussions}
}

\strong{Bug reporting}
\itemize{
\item \url{https://github.com/R-Lum/xlum/issues}
}

\strong{Project source code repository}
\itemize{
\item \url{https://github.com/R-Lum/xlum}
}

\strong{Package maintainer}

Sebastian Kreutzer, Geography & Earth Sciences, Aberystwyth University (United Kingdom),\cr
\email{sebastian.kreutzer@aber.ac.uk}

\strong{Funding}

Between 2020--2022, the work of Sebastian Kreutzer as maintainer of the package has received funding
from the European Union’s Horizon 2020 research and innovation programme under
the Marie Skłodowska-Curie grant agreement No 844457 (CREDit).
}
\keyword{package}
